Make ORDER BY pushdown optimization generic
This PR moves the `ORDER BY` and `LIMIT` pushdown optimization to a
separate module, with the intention to move this code to
[database-connector](https://github.com/duckdb/database-connector)
repo.
