+++ src/CMakeLists.txt
@@ -17,8 +17,11 @@ add_library(
   mysql_types.cpp
   mysql_utils.cpp)
 
-add_library(mysql_ext_dbconn OBJECT dbconn/query/query_writer.cpp
-                                    dbconn/table_scan/filter_pushdown.cpp)
+add_library(
+  mysql_ext_dbconn OBJECT
+  dbconn/query/query_writer.cpp dbconn/table_scan/filter_pushdown.cpp
+  dbconn/table_scan/filter_util.cpp
+  dbconn/optimizer/order_by_and_limit_optimizer.cpp)
 
 set(ALL_OBJECT_FILES
     ${ALL_OBJECT_FILES} $<TARGET_OBJECTS:mysql_ext_library>
