+++ src/include/dbconn/bind_data.hpp
@@ -1,0 +1,15 @@
+#pragma once
+
+#include "duckdb/function/function.hpp"
+
+#include "dbconn/optimizer/order_by_and_limit_bind_data.hpp"
+
+namespace dbconnector {
+
+class BindData : public duckdb::FunctionData {
+
+public:
+	virtual optimizer::OrderByAndLimitBindData &GetOrderByAndLimitBindData() = 0;
+};
+
+} // namespace dbconnector
