+++ src/include/dbconn/optimizer/order_by_and_limit_bind_data.hpp
@@ -1,0 +1,14 @@
+#pragma once
+
+#include <string>
+
+namespace dbconnector {
+namespace optimizer {
+
+struct OrderByAndLimitBindData {
+	std::string limit_clause;
+	std::string order_by_clause;
+};
+
+} // namespace optimizer
+} // namespace dbconnector
